backend/requirements.txt
Flask==2.1.1
pymongo==4.1.0
Flask-Cors==3.0.10
flask-socketio==5.3.3
Werkzeug==2.0.3
python-chess
